+++ scripts/equipment/consumables/poison-black-smear.js
@@ -3,13 +3,13 @@
     takes 4 (1d8) poison damage and is poisoned for 24 hours. While poisoned in this way, the creature smells of black
     smear. On a successful save, the creature takes half damage and isn't poisoned.
 */
-const version = "12.4.0";
+const version = "13.5.0";
 const optionName = "Black Smear Poison";
 const coatedName = "Black Smear Poisoned";
 const _flagGroup = "fvtt-trazzm-homebrew-5e";
 const _usesFlag = "coating-uses";
 const maxUses = 3; // for ammunition
-const saveDC = 11;
+const activityId = "poison-save";
 
 try {
     if (args[0].tag === "OnUse" && args[0].macroPass === "postDamageRoll") {
@@ -39,25 +39,28 @@ try {
             // request the saving throw
             let targetToken = workflow.hitTargets.first();
             if (targetToken) {
-                const hasResilience = HomebrewHelpers.hasResilience(targetToken.actor, "poison");
-                let saveRoll = await targetToken.actor.rollAbilitySave("con", {
-                    flavor: `${CONFIG.DND5E.abilities["con"].label} DC${saveDC} ${optionName}`,
-                    advantage : hasResilience
-                });
+                if (macroItem) {
+                    let activity = await macroItem.system.activities.find(a => a.identifier === activityId);
+                    if (activity) {
+                        const options = {
+                            midiOptions: {
+                                targetUuids: [targetToken.actor.uuid],
+                                noOnUseMacro: false,
+                                configureDialog: false,
+                                showFullCard: false,
+                                ignoreUserTargets: true,
+                                checkGMStatus: false,
+                                autoRollAttack: true,
+                                autoRollDamage: "always",
+                                fastForwardAttack: true,
+                                fastForwardDamage: true,
+                                workflowData: true
+                            }
+                        };
 
-                const saved = saveRoll.total >= saveDC;
-                let newDamageRolls = workflow.damageRolls;
-                let poisonRoll = undefined;
-
-                if (saved) {
-                    poisonRoll = await new CONFIG.Dice.DamageRoll('1d8 / 2', workflow.item.getRollData(), {type: 'poison'}).evaluate();
+                        await MidiQOL.completeActivityUse(activity, options, {}, {});
+                    }
                 }
-                else {
-                    poisonRoll = await new CONFIG.Dice.DamageRoll('1d8', workflow.item.getRollData(), {type: 'poison'}).evaluate();
-                    await HomebrewEffects.applyPoisonedEffect2024(targetToken.actor, macroItem, ['longRest'], 86400);
-                }
-                newDamageRolls.push(poisonRoll);
-                await workflow.setDamageRolls(newDamageRolls);
             }
         }
     }